create a cli folder
