+++ src/subliminal/cli/__init__.py
@@ -1,0 +1,6 @@
+"""CLI module."""
+
+from .cli import cli
+from .generate_config import generate_default_config
+
+__all__ = ['cli', 'generate_default_config']
