+++ src/subliminal/cli/commands/__init__.py
@@ -1,0 +1,6 @@
+"""CLI subcommands."""
+
+from .cache import cache, cache_file
+from .download_best import download
+
+__all__ = ['cache', 'cache_file', 'download']
